ai-worker/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
ollama==0.3.0
pydantic==2.7.4
pydantic-settings==2.3.4
structlog==24.2.0
prometheus-client==0.20.0
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
requests==2.32.3
asyncio==3.4.3
PyYAML==6.0.1